Renamed interface and implementations
