+++ src/main/java/com/g6jamm/stima/data/repository/mysql/ProjectRepositoryMySQLImpl.java
@@ -4,7 +4,7 @@ import com.g6jamm.stima.data.repository.ProjectRepository;
 import com.g6jamm.stima.data.repository.SubProjectRepository;
 import com.g6jamm.stima.data.repository.TaskRepository;
 import com.g6jamm.stima.data.repository.util.DbManager;
-import com.g6jamm.stima.domain.model.Project;
+import com.g6jamm.stima.domain.model.ProjectComposite;
 
 import java.sql.Date;
 import java.sql.PreparedStatement;
@@ -20,7 +20,7 @@ public class ProjectRepositoryMySQLImpl implements ProjectRepository {
     private final SubProjectRepository SUBPROJECT_REPOSITORY = new SubProjectRepositoryImpl();
 
     @Override
-    public Project createProject(Project project) {
+    public ProjectComposite createProject(ProjectComposite project) {
 
         try {
             String query =
@@ -42,7 +42,7 @@ public class ProjectRepositoryMySQLImpl implements ProjectRepository {
     }
 
     @Override
-    public Project getProject(int projectId) {
+    public ProjectComposite getProject(int projectId) {
 
         try {
             String query = "SELECT * FROM projects WHERE project_id = ? AND project_parent_id is NULL";
@@ -53,7 +53,7 @@ public class ProjectRepositoryMySQLImpl implements ProjectRepository {
             ResultSet rs = ps.executeQuery();
 
             if (rs.next()) {
-                return new Project.ProjectBuilder()
+                return new ProjectComposite.ProjectBuilder()
                         .projectId(projectId)
                         .projectName(rs.getString("name"))
                         .startDate(LocalDate.parse(rs.getString("start_date")))
@@ -90,7 +90,7 @@ public class ProjectRepositoryMySQLImpl implements ProjectRepository {
     }
 
     @Override
-    public void editProject(Project project) {
+    public void editProject(ProjectComposite project) {
 
         try {
             String query =
@@ -111,9 +111,9 @@ public class ProjectRepositoryMySQLImpl implements ProjectRepository {
     }
 
     @Override
-    public List<Project> getProjects() {
+    public List<ProjectComposite> getProjects() {
 
-        List<Project> projects = new ArrayList<>();
+        List<ProjectComposite> projects = new ArrayList<>();
         try {
 
             String query = "SELECT * FROM projects WHERE parent_project_id IS NULL";
@@ -122,8 +122,8 @@ public class ProjectRepositoryMySQLImpl implements ProjectRepository {
             ResultSet rs = ps.executeQuery();
 
             while (rs.next()) {
-                Project project =
-                        new Project.ProjectBuilder()
+                ProjectComposite project =
+                        new ProjectComposite.ProjectBuilder()
                                 .projectId(rs.getInt("project_id"))
                                 .projectName(rs.getString("name"))
                                 .startDate(LocalDate.parse(rs.getString("start_date")))
