+++ src/main/java/com/g6jamm/stima/data/repository/stub/ProjectRepositoryStub.java
@@ -1,8 +1,8 @@
 package com.g6jamm.stima.data.repository.stub;
 
 import com.g6jamm.stima.data.repository.ProjectRepository;
+import com.g6jamm.stima.domain.model.ProjectComposite;
 import com.g6jamm.stima.domain.model.Project;
-import com.g6jamm.stima.domain.model.ProjectComponent;
 import com.g6jamm.stima.domain.model.Task;
 
 import java.time.LocalDate;
@@ -11,7 +11,7 @@ import java.util.List;
 
 public class ProjectRepositoryStub implements ProjectRepository {
 
-  private static final List<Project> projects = new ArrayList<>();
+  private static final List<ProjectComposite> projects = new ArrayList<>();
 
   public ProjectRepositoryStub() {
     if (projects.isEmpty()) {
@@ -20,7 +20,7 @@ public class ProjectRepositoryStub implements ProjectRepository {
       SubProjectRepositoryStub subProjectRepositoryStub = new SubProjectRepositoryStub();
 
       projects.add(
-          new Project.ProjectBuilder()
+          new ProjectComposite.ProjectBuilder()
               .projectId(projects.size() + 1)
               .projectName("Projekt pink")
               .startDate(LocalDate.of(2021, 1, 1))
@@ -31,52 +31,52 @@ public class ProjectRepositoryStub implements ProjectRepository {
               .build());
 
       projects.add(
-          new Project.ProjectBuilder()
+          new ProjectComposite.ProjectBuilder()
               .projectId(projects.size() + 1)
               .projectName("Projekt lilla")
               .startDate(LocalDate.of(2021, 1, 1))
               .endDate(LocalDate.of(2021, 1, 2))
               .tasks(new ArrayList<Task>())
-              .subProjects(new ArrayList<ProjectComponent>())
+              .subProjects(new ArrayList<Project>())
               .colorCode("purple")
               .build());
 
       projects.add(
-          new Project.ProjectBuilder()
+          new ProjectComposite.ProjectBuilder()
               .projectId(projects.size() + 1)
               .projectName("Projekt grøn")
               .startDate(LocalDate.of(2021, 1, 1))
               .endDate(LocalDate.of(2021, 1, 2))
               .tasks(new ArrayList<Task>())
-              .subProjects(new ArrayList<ProjectComponent>())
+              .subProjects(new ArrayList<Project>())
               .colorCode("green")
               .build());
 
       projects.add(
-          new Project.ProjectBuilder()
+          new ProjectComposite.ProjectBuilder()
               .projectId(projects.size() + 1)
               .projectName("Projekt brun")
               .startDate(LocalDate.of(2021, 1, 1))
               .endDate(LocalDate.of(2021, 1, 2))
               .tasks(new ArrayList<Task>())
-              .subProjects(new ArrayList<ProjectComponent>())
+              .subProjects(new ArrayList<Project>())
               .colorCode("brown")
               .build());
     }
   }
 
   @Override
-  public Project createProject(Project project) {
+  public ProjectComposite createProject(ProjectComposite project) {
 
-    Project newProject =
-        new Project.ProjectBuilder()
+    ProjectComposite newProject =
+        new ProjectComposite.ProjectBuilder()
             .projectId(projects.size() + 1)
             .projectName(project.getName())
             .startDate(project.getStartDate())
             .endDate(project.getEndDate())
             .colorCode(project.getColorCode())
             .tasks(new ArrayList<Task>())
-            .subProjects(new ArrayList<ProjectComponent>())
+            .subProjects(new ArrayList<Project>())
             .build();
 
     projects.add(newProject);
@@ -84,11 +84,11 @@ public class ProjectRepositoryStub implements ProjectRepository {
   }
 
   @Override
-  public Project getProject(int projectId) {
+  public ProjectComposite getProject(int projectId) {
     return null;
   }
 
-  public List<Project> getProjects() {
+  public List<ProjectComposite> getProjects() {
     return projects;
   }
 
@@ -96,5 +96,5 @@ public class ProjectRepositoryStub implements ProjectRepository {
   public void deleteProject(int projectId) {}
 
   @Override
-  public void editProject(Project project) {}
+  public void editProject(ProjectComposite project) {}
 }
