+++ src/main/java/com/g6jamm/stima/domain/model/ProjectLeaf.java
@@ -3,7 +3,7 @@ package com.g6jamm.stima.domain.model;
 import java.time.LocalDate;
 import java.util.List;
 
-public class SubProject implements ProjectComponent {
+public class ProjectLeaf implements Project {
 
   /**
    * Model for subproject
@@ -18,7 +18,7 @@ public class SubProject implements ProjectComponent {
   private final LocalDate END_DATE;
   private final String COLOR_CODE;
 
-  public SubProject(SubProjectBuilder subProjectBuilder) {
+  public ProjectLeaf(SubProjectBuilder subProjectBuilder) {
     this.SUB_PROJECT_ID = subProjectBuilder.subProjectId;
     this.SUB_PROJECT_NAME = subProjectBuilder.subProjectName;
     this.SUB_PROJECT_TASKS = subProjectBuilder.subProjectTasks;
@@ -123,8 +123,8 @@ public class SubProject implements ProjectComponent {
       this.colorCode = null;
     }
 
-    public SubProject build() {
-      SubProject newSubProject = new SubProject(this);
+    public ProjectLeaf build() {
+      ProjectLeaf newSubProject = new ProjectLeaf(this);
       reset();
       return newSubProject;
     }
